requirements.txt
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
newsapi-python==0.2.6
psycopg2-binary==2.8.6
pycparser==2.20
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.23
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
